pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastmcp>=2.11.2",
    "markdownify>=1.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
    "google-genai>=1.60.0",
    "azure-cognitiveservices-vision-computervision>=0.9.1",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]
